Feat(violationRecord): Done implementing all the needed functions

+++ controllers/violationRecord.js
@@ -3,32 +3,38 @@ const violation = require('../models/violationRecord');
 const public = {
     //新增違規
     postViolationInsert: async (req, res) => {
-        violation.insertVR(req.body.vr_date, req.body.vr_type, 
-            req.body.resident_ID, req.body.housemaster_ID);
-        res.redirect('/violation/list');
+        await violation.insertVR(req.body.vr_type, req.body.resident_ID, req.session.user_ID);
+        res.redirect('/violationRecord/list');
     },
 
-    //刪除違規紀錄
-    postViolationDelete: async (req, res) => {
-        violation.deleteVR(req.body.vr_ID);
-        res.redirect('/violation/list');
-    },
-
-    //修改違規紀錄
+    //修改或刪除違規紀錄
     postViolationModify: async (req, res) => {
-        violation.modifyVR(req.body.vr_ID, req.body.vr_date, req.body.vr_type);
-        res.redirect('/violation/list');
+        if (req.body.delete) {
+            await violation.deleteVR(req.body.vr_ID);
+		} else {
+            await violation.modifyVR(req.body.vr_ID, req.body.vr_type);
+		}
+        res.redirect('/violationRecord/list');
     },
 
     //查看某住宿生違規紀錄
     getViolationResident: async (req, res) => {
-        const violationResident = await violation.selectResidentVR(req.body.resident_ID);
-        res.renderInjected('violationRecord', {violationResident});
+        const violationResident = await violation.selectResidentVR(req.params.user_ID);
+        console.log(violationResident);
+        res.renderInjected('violationRecord', {
+            violationInfos: violationResident
+        });
     },
 
     //查看所有違規紀錄
     getViolationList: async (req, res) => {
-        const violationInfos = await violation.selectAllVR();
+        if(req.session.role == 'resident_student'){
+            res.redirect(`/violationRecord/${req.session.user_ID}`);
+        }
+        violationInfos = await violation.selectAllVR();
+        if(req.session.role == 'housemaster'){
+            violationInfos = violationInfos.filter(element => element.housemaster_ID == req.session.user_ID);
+        }
         res.renderInjected('violationRecord', {violationInfos});
     },
